coggrinder.console.ConsoleTestUtil: - Updated to use the new GoogleServiceTask/TaskListService classes.
coggrinder.services.tasktree_services.TaskTreeService:
- Updated to use the new GoogleServiceTask/TaskListService classes.
.TaskTreeServiceTestCommon:
- Updated to use the new InMemoryTask/TaskListService classes.

coggrinder.services.task_services:
- Major refactor that resulted in the creation of a pair of implementing
classes for the Task and TaskList services. Each service now has an
implementation that uses the remote Google GTask Services as its data
store, and another that uses a local, in-memory dictionary as the data
store.
- The in-memory services currently serve primarily as testing/debug
support classes, but could easily be used to provide a local task data
store (a simple option would be to pickle/unpickle the dict objects) in
the future.
- Added various other Error and common architecture classes to support
the refactor.