TreeNode.path argument is now dynamic and calculated at access-time.
TreeNode.path is switched from being a static property that is set to
one that is calculated at the time of being accessed. This is done so
that when the parent node's path is updated (most likely due to a tree
sort operation) the change in position is automatically pushed down to
the descendant nodes. The implementation of .path is currently combining
the parent node's path with the index of the tree node in the parent's
children.