Updates convenience method for creating test task data titles.
Updates the TestDataEntitySupport.create_full_title method to support
creating titles (and therefore, entity IDs) from an arbitrary number of
short title subsections. This method makes it easier to generate
titles/ids for deeply nested test data entities like a Task A-C-B-D.

Swapped the ordering of the arguments of create_full_title to allow the
short title to be a variable length argument.