Adds task_data init argument support to TaskTree.
TaskTree gains support for providing task data as a single, flat
dictionary (as opposed to the dual dictionaries of tasklists and
all_tasks, the latter of which is not flat). Flat collections of all
task data should be easier to build than separate and more complicated
dictionaries that split TaskList- and Task-type entities.