Moves test task data entities into tasks module.
Moves the test task data convenience entities from the
coggrinder.entities.tasktree module to coggrinder.entities.tasks. This
is intended to trim the size of the tasktree module and make it a bit
easier to navigate around. Additionally, the test task data entities
were more appropriately placed in the tasks module. Updates references
to the moved classes where needed.