Adds TaskDataSorter utility class.
Adds TaskDataSorter, a utility class that sorts an unordered, flat
list of task data entities into an ordered list that reflects the
in-order traversal of the task tree the entities represent.

Adds test coverage for the TaskDataSorter methods.

Updates the str/repr results of BaseTaskEntity (and therefore, TaskList
and Task) to create a string bounded on each end with angle brackets,
and with property-value pairs delimited by semi-colons.

Updates TestDataTaskList and TestDataTask init methods to accept
variable-length short title sections in their arguments.