Fixes Task deletion, descendant Task survival bug.
Fixes a bug where deleting a Task also recursively deregistered all
descendant Tasks, which effectively deleted those Tasks from the
TaskTree. Note that this descendant deregistration behavior is still
present and appropriate for a TaskList deletion, as that operation also
renders all Tasks belonging to the TaskList as unavailable.

Adds test coverage to verify that this bug is fixed.

Updates the test fixture setup of test case PopulatedTaskTreeTest to
make use a clone of the expected TaskTree for a working TaskTree.