Changes the way tree view state is maintained.
Instead of trying to rebuild tree view state (which rows/entities are
expanded and/or selected), updates the code to simply rebuild the tree
store after every update. Building the treestore from a tasktree is
much simpler than trying to track and selectively update the treestore
as the tasktree changes. 


To compensate, the tree view stat