Updates the entryExit log decorator to support more output options.
Updates the entryExit log decorator to support two new output options:
the ability to record the arguments the wrapped method was called with
and the result of the wrapped method call.

Reworks the simple test process to ensure that the decorator handles
instance attributes appropriately, and to exercise the new output
options.