Adds EventRegistry and supporting tests.
Adds the new EventRegistry, a simple system for propagating events.
EventRegistry collects event notifications from notifiers and can then
propagate the event information out to a collection of listeners that
are registered to handle that specific event's name. 

Adds a pair of initial tests under a single test case to exercise the
basic functionality of the EventRegistry.