Adds new eventAware class decorator.
EventAware is a class decorator that uses introspection to register
methods in the class that are designated as event listeners (handlers).
This works in tandem with the eventListener method decorator, which
serves as the marking designator and collects information about the
event the decorated method will be listening for. Currently this is
limited to event name, but could be expanded in the future.

Adds a more informative str representation to EventInfo.

Adds new deregister_event_listeners method to the EventRegistry, which
facilitates unregistering all event listeners for a particular event. 

Updates the test cases to clear the event registry event handlers list
before and after each test. This helps prevent the statically-stored
event registry information from cross-contaminating classes. Adds new
test to verify this functionality.

Add new test case to test the eventAware class decorator.