Updates widget event handling config to use the new event system.
Alters the init signature of BinaryIconButton. The initialization now
requires the provision of a click event name, which is used to register
the button click event with the new event system.

Updates "remove" Task/List mentions to read "delete", so that they
better reflect the nomenclature used elsewhere in the system (in
particular, the services layer).

Adds TaskTreeEvents, a simple class that collects the event name
constants used within this module.

Adds eventListener and eventAware decorators to _handle_revert_event
and TaskTreeWindowController, respectively. These decorations work
together to hook the handling method up to notifications for the revert
action.

Cleans out some old, unused imports.