Updates TreeState default values.
Changes the name of TreeState to TaskTreeViewState. 

Converts the default init values from False to None. This change adds a
third, "undefined" option to the tree state properties, allowing for
TaskTreeViewStates to be configured to reflect a property that should be
updated (True, False) or left as-is (None).

Moves both TaskTreeViewState and EmptyTreeStateError to be classes
directly belonging to the model, rather than inner classes of
TaskTreeViewController. This change anticipates usage of the classes
outside of the TaskTreeViewController (specifically, they are expected
to be used by the TaskTreeWindowController).