Adds new add_child method to TreeNode.
Adds the new add_child method to TreeNode, which is effectively what
attach_to_parent previously handled, but with the arguments swapped.
Updates attach_to_parent to now use add_child from the supplied parent
argument. This change was made because adding a child seems to be more
intuitive than attaching to a parent.