Updates entity service determination procedure.
Updates the procedure for determining the proper service (Task or
TaskList) to use with an entity. The new method relies upon the presence
of a task_status property on the entity (Tasks have one, TaskLists
don't). The update also separates out this determination process so it
can be reused and updates push deleted and push added to use the new
method.