Updates handling for save, refresh events.
Updates the handling for the save and refresh events. Refreshing now
wipes away the local information, replacing it with "fresh" information
from the task data services (the Google Task services in normal
operation). Saving now asks the TaskTreeService to push the local
updates up to the server before then refreshing the local data with
updated remote information and updating the view.

Note that TaskTreeService currently does not fully implement the
push_task_data operation, and therefore this is a UI-only change at the
moment.