Updates task data entity ID handling.
Updates entity ID handling to use a set of three properties: entity_id,
local_id, and persistence_id. local_id is assigned to any new entities
that are created without an explicit entity ID argument, and is built
from a UUID generator. persistence_id is the ID that will be used to
identify the entity within the persistence layer (for instance, for
GoogleTaskListService, this would map to the "id" JSON attribute).
persistence_id is None for locally created entities until they are
persisted. entity_id is a property that points to either the local_id or
persistence_id, depending on circumstances. Setting the entity_id always
sets the persistence_id of the entity. Getting the entity_id will
return the persistence_id, unless undefined, in which case the local_id
(which is always populated) will be returned. This scheme is developed
to allow the system to quickly generate local handle IDs for entities
while still maintaining the separate ID system necessary to interface
with the persistence layer.