Adds treeless_value property to TreeNode.
Adds treeless_value property to TreeNode. This property is meant to
provide a value for the TreeNode without taking into consideration the
parent and child references the TreeNode maintains. Enabling this
separation is needed to update add_child such that is will allow for
comparison between rich subclasses such as TaskList and Task.

Removes check in Tree insert_node to ensure a single root node, as the
Tree is now its own root node.

Removes the Tree.remove method. This was not being used by any client
component.

Removes tests that covered now-deprecated use cases.