Adds support for the Task tasklist_id property.
Adds support for the Task tasklist_id property, including configuration
via the init arguments. If the tasklist_id is not specified, but a
parent of the Task is a TaskList, retrieving the tasklist_id will set
the value of the property to be the ID of the parent TaskList. Adds a
supporting test.

Makes a small change to the sorted add_child implementation to solve a
bug that was causing the normal/full value of a new child to be compared
with the treeless values of siblings (in both cases, the values used
should be treeless).