Updates title to position value conversion.
Updates the conversion of the TestDataGoogleServicesTask title to
position value by changing the process from operating only on the short
title sections to being able to operate on the full title of the entity.
This makes the algorithm a bit more versatile and allows the continued
use of previously written code against this class without modification.