Updates the TaskTree tree builder algorithm.
Updates the algorithm used by TaskTree to be much simpler and more
direct. Instead of recursing through the task data, this method simply
loops through all entities in the task data set and links them to their
parent. This also enables passing in task data sets that lack defined
parent/child relationships between the entities (such as would be the
case if a set of task data was generated with TreeNode.clean_clone
output).

Updates the TaskTreeComparator to compare entities using the entities'
treeless values where searching for updated entities.