Updates InMemory services to use clean entity clones.
Updates the InMemoryTaskService and InMemoryTaskListService service
implementations to return clean entity clones. This prevents clients
from directly modifying the services' entity stores. This change allows
the two services to better imitate their Google Services counterparts,
which would also be returning "detached" entities in similar
circumstances.

Updates tests to account for and verify changes to the InMemory services
and the addition of the tasklist_id property to Task entities.