Updates Task parent_id and tasklist_id properties.
Updates the Task parent_id and tasklist_id properties to defer to the
parent and tasklist references when accessing those IDs. This solves a
bug where the parent/tasklist was persisted, received a new
(persistence) ID, and the child Task was unaware of the change (and
therefore persisted with an outdated parent ID).

Adds update_from_str_dict method which allows Task/List instances to
have their attributes updated in-place with the results of a service
operation, rather than replacing the entire instance. Adds test to
verify the new functionality.