Fixes bug in adding a parent-child chain of task data entities.
Fixes a bug that prevented the persistence of a sequence of added
entities if those entities formed a parent-child chain (or "task
branch"). Entities are now added in order from shallowest to deepest,
and within groups the ordering between the sibling entities is the same
order they will be added through the insert service operations. Adds new
test to verify functionality.