Enables and updates unit tests.
Enables and updates the TaskTreeReorderTest tests. These tests cover
Task reordering (both up and down), as well as Task promotion and
demotion. The tests are updated to reflect changes in the client usage
patterns that have been the driver and result of (task data) refactoring
effots in other components.

Also enables the TaskTreeSortTest.

Fixes an error in the find_updated_ids algorithm where an entity would
be declared "updated" if it had an updated updated_date property

Enables and updates the TaskTreeComparatorFindUpdatedTest to align with
current use cases and verify that the updated TaskTreeComparator find
updated algorithm is working properly.

Splits off the tests involving determining which entities have been
reordered into a separate set of tests, which are currently disabled.