Updates find_moved_task_ids algorithm to use previous_task_id.
Updates the fine_moved_task_ids algorithm to use the new Task
previous_task_id attribute. This fixes a bug that was caused by
comparisons based on the child_index value, which could change even if
the Task did not need to be the subject of a move operation (it still
had the same proceeding Task). Updates an existing test to verify the
behavior.